+++ data/tests/code-in-pre.js
@@ -1,0 +1,18 @@
+docTests.codeInPre = {
+  name: "code_in_pre",
+  desc: "code_in_pre_desc",
+  check: function checkCodeInPre(rootElement) {
+    var codesInPres = rootElement.querySelectorAll("pre code");
+    var matches = [];
+
+    for (var i = 0; i < codesInPres.length; i++) {
+      matches.push({
+        msg: codesInPres[i].outerHTML
+      });
+    }
+
+    return matches;
+  },
+  type: ERROR,
+  count: 0
+};
