+++ data/tests/testlist.js
@@ -1,0 +1,25 @@
+// This will also be the ordering of the tests in the sidebar
+
+exports.testList = [
+  "old-urls.js",
+  "empty-elements.js",
+  "languages-macro.js",
+  "empty-brackets.js",
+  "style-attribute.js",
+  "name-attribute.js",
+  "span-count.js",
+  "pre-without-class.js",
+  "summary-heading.js",
+  "jsref-with-params.js",
+  "example-colon-heading.js",
+  "alert-print-in-code.js",
+  "html-comments.js",
+  "font-elements.js",
+  "http-links.js",
+  "macro-syntax-error.js",
+  "wrong-highlighted-line.js",
+  "api-syntax-headlines.js",
+  "wrong-syntax-class.js",
+  "code-in-pre.js",
+  "line-length-in-pre.js",
+];
