+++ lib/main.js
@@ -2,7 +2,7 @@ const data = require("sdk/self").data;
 const tabs = require("sdk/tabs");
 const localize = require("sdk/l10n").get;
 const prefs = require('sdk/simple-prefs').prefs;
-
+const testList = require("../data/tests/testlist").testList;
 
 var sidebar = require("sdk/ui/sidebar").Sidebar({
   id: 'mdn-doc-tests',
@@ -11,7 +11,11 @@ var sidebar = require("sdk/ui/sidebar").Sidebar({
   onReady: function (sbWorker) {
     sbWorker.port.on("runTests", function() {
       tabWorker = tabs.activeTab.attach({
-        contentScriptFile: [data.url("doctests.js"), data.url("runtests.js")]
+        contentScriptFile: [
+          "./doctests.js",
+          ...testList.map(test => "./tests/" + test),
+          "./runtests.js"
+        ]
       });
       tabWorker.port.emit("runTests");
       tabWorker.port.on("test", function(testObj, id){
